Proposed ArrayAccess wrapper. - moved to Helper namespace - removed __call function from wrapper - updated test
